requirements.txt
streamlit
requests
python-dotenv
PyPDF2
python-docx
scikit-learn
numpy
pandas
